file: Supress diff on create
Supress output of content diff on create - in case when the file is not
present on filesystem yet.

When large script files are created by remote_file or template resource
the current behavior consumes a lot of log lines which is noisy.

Original itamae behaves the same [1]. The current behavior has been
preserved since initial implementation [2] and I believe this is not
intentional.

[1]: https://github.com/itamae-kitchen/itamae/blob/fd48cb50fd63777139c2f62cb809fb16812c945a/lib/itamae/resource/file.rb#L63
[2]: https://github.com/itamae-kitchen/mitamae/commit/65df89cee0882a5892cf197e94787b3f1a2df214